Move things in hir_ty into submodules
 - all the types that will be replaced by Chalk go to `types`
 - `TypeWalk` impls go to `walk`
